Initiate Scaffold for User Ratings
I initiated the scaffold command for my user ratings.  This provided me
with the updated routes, generated a controller and a model.  I have set
up this scaffold to reference both the user and the album table id.

On the user_ratings, I need to :
* Add testing
* Update the controllers, routes, and models

Overall:
I need to look into the links between the tables and establishing the
proper relationship.