Commit Updates to Routes and Controllers
Update to album routes and and user ratings routes and controller.
For albums, the ability to modify or delete an album record was removed.
For user ratings, route to find the current user's rating for an album
was updated to be more inline with REST-ful convention.  Improved
security to only allow the current user to only be able to delete or
update their own ratings.